refactor: booking폴더 하위의 pages폴더 제거 #82
